# Speed up _match function [run_process_replay]