@geohot 2024-06-19
Minor speedup for more lines/complexity isn't worth it. The whole algorithm is wrong.